requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.3
behave==1.2.6
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
colorama==0.4.6
cssselect==1.3.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
execnet==2.1.1
fasteners==0.19
filelock==3.18.0
h11==0.14.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
MouseInfo==0.1.3
mycdp==1.1.1
numpy==2.2.4
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
parse==1.20.2
parse_type==0.6.4
pdbp==1.7.0
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
psutil==7.0.0
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==1.0.9
pynose==1.5.4
pyotp==2.9.0
pyperclip==1.9.0
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
pytest==8.3.5
pytest-html==4.0.2
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==15.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
sbvirtualdisplay==1.4.0
selenium==4.30.0
seleniumbase==4.36.2
setuptools==80.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tabcompleter==1.4.0
tomli==2.2.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
wsproto==1.2.0
